Configuring XDebug
Confirmed to work with PHPStorm 10 (yay!).

These configuration settings do add some security risks - but then this is a debugging extension so should not be running in publicly accessible environments, or any environment with sensitive information.

The Ansible INI module used to set these options is *very* inflexible as to who INI files are structured. Fortunately PHP is not when reading them so the modifications we make to add a section header are essentially ignored by PHP.